+++ core/src/main/java/com/graphhopper/routing/util/parsers/OSMRoadAccessParser.java
@@ -79,6 +79,8 @@ public class OSMRoadAccessParser implements TagParser {
                 return Arrays.asList("hgv", "motor_vehicle", "vehicle", "access");
             case PSV:
                 return Arrays.asList("psv", "motor_vehicle", "vehicle", "access");
+            case BUS:
+                return Arrays.asList("bus", "psv", "motor_vehicle", "vehicle", "access");
             default:
                 throw new IllegalArgumentException("Cannot convert TransportationMode " + mode + " to list of restrictions");
         }