+++ core/src/test/java/com/graphhopper/routing/util/TagParsingTest.java
@@ -129,37 +129,6 @@ class TagParsingTest {
     }
 
     @Test
-    public void testCompatibilityBug() {
-        EncodingManager manager2 = EncodingManager.create("mtb");
-        ReaderWay osmWay = new ReaderWay(1);
-        osmWay.setTag("highway", "footway");
-        osmWay.setTag("name", "test");
-
-        MountainBikeTagParser parser = new MountainBikeTagParser(manager2, new PMap());
-        parser.init(new DateRangeParser());
-        IntsRef flags = parser.handleWayTags(manager2.createEdgeFlags(), osmWay);
-        double singleSpeed = parser.avgSpeedEnc.getDecimal(false, flags);
-        assertEquals(BikeCommonTagParser.PUSHING_SECTION_SPEED, singleSpeed, 1e-3);
-        assertEquals(singleSpeed, parser.avgSpeedEnc.getDecimal(true, flags), 1e-3);
-
-        EncodingManager manager = EncodingManager.create("mtb,bike,foot");
-        FootTagParser footParser = new FootTagParser(manager, new PMap());
-        footParser.init(new DateRangeParser());
-        MountainBikeTagParser bikeParser = new MountainBikeTagParser(manager, new PMap());
-        bikeParser.init(new DateRangeParser());
-
-        flags = footParser.handleWayTags(manager.createEdgeFlags(), osmWay);
-        flags = bikeParser.handleWayTags(flags, osmWay);
-        DecimalEncodedValue bikeSpeedEnc = manager.getDecimalEncodedValue(VehicleSpeed.key("mtb"));
-        assertEquals(singleSpeed, bikeSpeedEnc.getDecimal(false, flags), 1e-2);
-        assertEquals(singleSpeed, bikeSpeedEnc.getDecimal(true, flags), 1e-2);
-
-        DecimalEncodedValue footSpeedEnc = manager.getDecimalEncodedValue(VehicleSpeed.key("foot"));
-        assertEquals(5, footSpeedEnc.getDecimal(false, flags), 1e-2);
-        assertEquals(5, footSpeedEnc.getDecimal(true, flags), 1e-2);
-    }
-
-    @Test
     public void testSharedEncodedValues() {
         BooleanEncodedValue carAccessEnc = VehicleAccess.create("car");
         BooleanEncodedValue footAccessEnc = VehicleAccess.create("foot");